c2-telegram/requirements.txt
requests==2.31.0
PyYAML==6.0
redis==5.0.1